Fix Coverity defect
CID 456641:    (PRINTF_ARGS)
Argument "saml_port" to format specifier "%d" was expected to have
type "int" but has type "long".

+++ src/config.h
@@ -92,7 +92,7 @@ struct vpn_config {
 	int			password_set;
 	char			otp[OTP_SIZE + 1];
 	char			*cookie;
-	int			saml_port;
+	uint16_t		saml_port;
 	char			saml_session_id[MAX_SAML_SESSION_ID_LENGTH + 1];
 	char			*otp_prompt;
 	unsigned int		otp_delay;
